Refactor: Removed assumption about the state representation.
